requirements.txt
# Core LangChain & LangGraph
langgraph==0.2.34
langchain==0.3.1
langchain-groq==0.2.0
langsmith==0.1.129

# Document Generation
python-docx==1.1.2
Pillow==10.4.0

# Code Execution & Validation
docker==7.1.0
pylint==3.3.1
flake8==7.1.1
black==24.10.0

# Security Scanning
bandit==1.7.10

# Database & State Management
redis==5.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
slowapi
python-jose[cryptography] 
bcrypt==4.0.1 
passlib[bcrypt]
alembic

# API & Web Framework
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
tenacity==8.5.0

# Monitoring & Logging
structlog==24.4.0
python-docx
watchdog

#uvicorn app:app --reload